requirements.txt
Django==5.0.6
Pillow==10.4.0
psycopg2-binary==2.9.9
pandas
openpyxl
xlrd
pdfplumber
python-docx
python-dotenv
